📈(xiti) migrate to piano analytics
ATInternet will deprecate soon its Analytics Suite 2 in favor of Piano
Analytics. Currently, to send analytic events we are using smartTag SDK
which is able to send data to Piano Analytics but not in an optimal way.
So we take the opportunity of the deprecation of AS2 to switch to the
new Piano Analytics SDK to send analytics event.