♿️(ademe) use lighter color for empty course detail
When no opened course runs are available, a label is displayed to explain that
to the user but currently the shade difference between the color label and its
background does not comply with RGAA.

It appears that currently a light color on a tealish background does not comply
at all with RGAA! The only color that could be used is black. But currently, all
text in the course subheader is white so if we use a black color for this label
only we are breaking ui consistency.

In other hand, the new constrast color norm used by the WCAG next version 3,
APCA, challenges RGAA in this case. It appears that a pale-grey 16px bold text
on a tealish background could comply with APCA. So as a take-off, in order to
improve accessibility and keep ui consistency, we could use a 'pale-grey' color
for this label.

Related issue : #27