👷(circleci) create canary workflow
We want to publish every day a canary version of every sites. To do
that, a canary workflow is always present in the config.yml file and is
generated from the `bin/ci update` command.
The hub-canary job updates the richie-education npm package to the next
version. We can't commit this because it will always be the same version
in the yarn.lock file.
A new stage canary is added in the Dockerfile. It removes richie package
and install it again using the --pre option with pip.