🩹️(funmooc) fix type issue that broke canary build
The next version of `richie-education` package rename Enrollment type into
UnknownEnrollment. As we were using this type in an overridden module, tsc
command raises an error and prevent the frontend to be built.
In order to be able to build canary images without requiring to release a new
version of Richie, we simply remove the use of the problematic type. As the
impacted module is only reserved for development purpose, this workaround looks
acceptable.