🚩(funmooc) split REACT_DASHBOARD feature flag
We would like to be able to enable the new dashboard but without updating user
menu entries. In order to do that we split `REACT_DASHBOARD` feature flag in
two: `REACT_DASHBOARD` which, if it is True, add dashboard urls in Django urls.
`REACT_DASHBOARD_MENU_LINKS` which, if it is True, add new dashboard links into
user menu.