🔧(requirements) add django-cms fork into requirements
Currently, we are installing our django-cms fork through a command ran into the
Dockerfile. Since an update of Docker, we had recently to update this command
but this introduced a bug which leads to install Django 5.x when installing our
fork. In order to keep django version constraint, we decide to add our fork into
 each base.txt requirements file of each our sites.