🐛(canary) collector step must be made with the richie next version
The collector step in the Dockerfile is made using the richie version
set in the backend requirements. If new statics are added in the next
tag then they will not be present. To fix that, the uninstallation of
richie and the installtion of the next version is made in the core
stage. Doing that, the canary stage is not needed anymore and the
argument BUILD_NEXT_FRONT is rename in BUILD_NEXT_VERSION to be more
accurate.